fix: improve UI for tool execution failed
