+++ examples/ui/frontend/src/components/events/tool-error.tsx
@@ -1,6 +1,8 @@
 import React, { JSX, useState } from "react";
 import { ChevronRight, AlertTriangle, AlertCircle } from "lucide-react";
 import { PLATFORM_MODE } from "@/lib/config";
+import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
+import { vscDarkPlus } from "react-syntax-highlighter/dist/esm/styles/prism";
 
 interface ToolErrorEventProps {
   payload?: {
@@ -34,9 +36,46 @@ const ToolErrorEvent: React.FC<ToolErrorEventProps> = ({ payload, openUpgradeMod
     return `${toolName} failed: ${truncatedError}`;
   };
 
+  const isExecuteScript = payload.tool_name?.toLowerCase().includes("execute script") || 
+                         payload.tool_name?.toLowerCase().includes("script");
+
+  const getLanguage = (language?: string): string => {
+    if (!language) return "javascript";
+    
+    const languageMap: Record<string, string> = {
+      "python": "python",
+      "javascript": "javascript",
+      "js": "javascript",
+      "typescript": "typescript",
+      "ts": "typescript",
+      "bash": "bash",
+      "shell": "bash",
+      "sh": "bash",
+      "php": "php",
+      "java": "java",
+      "cpp": "cpp",
+      "c++": "cpp",
+      "c": "c",
+      "go": "go",
+      "rust": "rust",
+      "ruby": "ruby",
+      "sql": "sql",
+      "html": "html",
+      "css": "css",
+      "json": "json",
+      "yaml": "yaml",
+      "yml": "yaml",
+      "markdown": "markdown",
+      "md": "markdown"
+    };
+    
+    return languageMap[language.toLowerCase()] || "javascript";
+  };
+
   const renderExpandedContent = (): JSX.Element => {
     const { tool_name, input_params, error } = payload;
 
+    // Default error display for non-script tools
     return (
       <div className="mx-3 mb-4 bg-red-50 border border-red-200 rounded-md overflow-hidden">
         <div className="flex items-center px-3 py-2 bg-red-100 border-b border-red-200">
@@ -67,6 +106,117 @@ const ToolErrorEvent: React.FC<ToolErrorEventProps> = ({ payload, openUpgradeMod
 
   const isUpgradeError = payload.isUpgradeErrorMessage;
 
+  // For script errors, display directly without dropdown
+  if (isExecuteScript && payload.input_params) {
+    const { tool_name, input_params, error } = payload;
+    const code = input_params.code as string || input_params.script as string || "Unknown script";
+    const language = getLanguage(input_params.language as string);
+
+    return (
+      <>
+        <div className="flex justify-start">
+          <div className="flex items-center space-x-2 px-3 py-2">
+            <AlertTriangle className="w-3 h-3 text-red-600" />
+            <span className="text-xs text-gray-500 truncate max-w-md">
+              <strong>{getDisplayContent()}</strong>
+            </span>
+            <button
+              onClick={toggleExpanded}
+              className="flex items-center py-0.5 text-xs text-gray-400 hover:text-gray-600 transition-colors"
+              title={isExpanded ? "Hide details" : "Show details"}
+            >
+              <div
+                className={`transition-transform duration-200 ${
+                  isExpanded ? "rotate-90" : "rotate-0"
+                }`}
+              >
+                <ChevronRight className="w-3 h-3" />
+              </div>
+            </button>
+          </div>
+        </div>
+
+        <div
+          className={`grid transition-all duration-300 ease-in-out ${
+            isExpanded
+              ? "grid-rows-[1fr] opacity-100"
+              : "grid-rows-[0fr] opacity-0"
+          }`}
+        >
+          <div className="overflow-hidden">
+            <div className="mx-3 mb-4 bg-gray-900 text-white rounded-md overflow-hidden border-red-200">
+              <div className="p-3">
+                <div className="mb-3">
+                  <div className="text-xs text-gray-400 mb-1">Code:</div>
+                  <div 
+                    className="max-h-64 overflow-y-auto"
+                    style={{
+                      scrollbarWidth: 'thin',
+                      scrollbarColor: '#4B5563 #1F2937'
+                    }}
+                  >
+                    <SyntaxHighlighter
+                      language={language}
+                      style={vscDarkPlus}
+                      customStyle={{
+                        margin: 0,
+                        borderRadius: "0.375rem",
+                        fontSize: "0.875rem",
+                      }}
+                      showLineNumbers
+                    >
+                      {code}
+                    </SyntaxHighlighter>
+                  </div>
+                </div>
+                {error && (
+                  <div>
+                    <div className="text-xs text-gray-400 mb-1">Error:</div>
+                    <div 
+                      className="bg-gray-800 p-3 rounded-md font-mono text-sm text-gray-300 whitespace-pre-wrap break-words max-h-64 overflow-y-auto"
+                      style={{
+                        scrollbarWidth: 'thin',
+                        scrollbarColor: '#4B5563 #1F2937'
+                      }}
+                    >
+                      {String(error)}
+                    </div>
+                  </div>
+                )}
+              </div>
+            </div>
+          </div>
+        </div>
+
+        {/* Upgrade modal trigger for upgrade errors */}
+        {isUpgradeError && (
+          <div className="mx-3 mb-2">
+            {!PLATFORM_MODE ? (
+              <button
+                onClick={() => window.open('https://agi.pandas-ai.com/upgrade', '_blank', 'noopener,noreferrer')}
+                className="inline-flex items-center px-3 py-1.5 text-xs font-medium text-orange-700 bg-orange-100 border border-orange-300 rounded-md hover:bg-orange-200 transition-colors"
+              >
+                <AlertCircle className="w-3 h-3 mr-1" />
+                Upgrade Required
+              </button>
+            ) : (
+              openUpgradeModal && (
+                <button
+                  onClick={openUpgradeModal}
+                  className="inline-flex items-center px-3 py-1.5 text-xs font-medium text-orange-700 bg-orange-100 border border-orange-300 rounded-md hover:bg-orange-200 transition-colors"
+                >
+                  <AlertCircle className="w-3 h-3 mr-1" />
+                  Upgrade Required
+                </button>
+              )
+            )}
+          </div>
+        )}
+      </>
+    );
+  }
+
+  // For non-script errors, use the dropdown interface
   return (
     <>
       <div className="flex justify-start">
